RAG/backend/requirements.txt
# ===========================================
# OnsenRAG - 必要ライブラリ一覧
# ===========================================
# インストール: pip install -r requirements.txt

# --- LangChainフレームワーク ---
# RAGの構築に使用する主要フレームワーク
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
# テキスト分割（RecursiveCharacterTextSplitter, from_tiktoken_encoder）
langchain-text-splitters>=0.2.0

# --- ベクトルDB ---
# Chroma: 無料でローカルで動作するベクトルデータベース
chromadb>=0.4.0

# --- Embeddingモデル ---
# sentence-transformers: 無料で使えるテキスト埋め込みモデル
sentence-transformers>=2.2.0

# --- LLM API ---
# Google Gemini: 無料枠あり・高性能・日本語対応（推奨）
langchain-google-genai>=4.0.0
# OpenAI: GPTモデルを利用するためのクライアント
openai>=1.0.0
# Groq: 無料・高速LLM推論（Llama 3.3等）
langchain-groq>=0.1.0

# --- ドキュメント読み込み ---
# pypdf: PDF文書の読み込みに使用
pypdf>=3.0.0

# --- 環境変数管理 ---
# python-dotenv: .envファイルからAPIキーなどを読み込む
python-dotenv>=1.0.0

# --- ハイブリッド検索用 ---
# rank_bm25: BM25キーワード検索アルゴリズム
rank_bm25>=0.2.2

# --- 日本語テキスト処理（オプション） ---
# tiktoken: OpenAIのトークン計算用
tiktoken>=0.5.0

# --- Web API（チャットUI用） ---
# FastAPI: 高速なPython Web APIフレームワーク
fastapi>=0.100.0
# uvicorn: ASGI対応Webサーバー（FastAPIの実行に必要）
uvicorn[standard]>=0.20.0
# gunicorn: Cloud Run 本番用ワーカー管理
gunicorn>=21.0.0

# --- Firebase / GCP ---
# Firebase Admin SDK: Firestore への接続・データ保存
firebase-admin>=6.0.0
